refresh

+++ app/Http/Controllers/Htmx/HTMXArticleController.php
@@ -19,14 +19,16 @@ class HTMXArticleController extends Controller
         if (auth()->check()) {
             $isArticleFavoritedByUser = $article->favoritedByUser(auth()->user());
         }
-       
 
         return view('articles.partials.show', [
             'article' => $article,
             'favorite_count' => $article->favoritedUsers->count(),
             'is_favorited' => $isArticleFavoritedByUser
+        ])
+        .view('components.navbar', ['navbar_active' => ''])
+        .view('components.htmx.head', [
+            'page_title' => Str::words($article->title, 40, '') . ' —'
         ]);
-        
     }
 
     public function favorite(Article $article)
@@ -96,7 +98,9 @@ class HTMXArticleController extends Controller
         if (auth()->guest()) {
             return Helpers::redirectToSignIn();
         }
-
+        if(auth()->user()->role<2){
+            return Helpers::redirectToHome();
+        }
         $article->delete();
 
         return response()->view('components.redirect', [
